requirements.txt
python-telegram-bot==21.7
yfinance==0.2.54
pandas==2.2.3
python-dotenv==1.0.1
